+++ packages/featureflags/src/index.ts
@@ -1,0 +1,1 @@
+export { default, useFeatureFlags } from './react'